+++ scripts/test.json
@@ -1,0 +1,1 @@
+{'model': 'meta/llama-3_1-8b-instruct', 'messages': [{'role': 'user', 'content': 'Hello! How are you?'}, {'role': 'assistant', 'content': 'Hi! I am quite well, how can I help you today?'}, {'role': 'user', 'content': 'Can you get me weather for San Francisco, CA?'}], 'top_p': 1, 'n': 1, 'max_tokens': 200, 'stream': False, 'tools': [{'type': 'function', 'function': {'name': 'get_current_weather', 'description': 'Get the current weather', 'parameters': {'type': 'object', 'properties': {'location': {'type': 'string', 'description': 'The city and state, e.g. San Francisco, CA'}, 'format': {'type': 'string', 'enum': ['celsius', 'fahrenheit'], 'description': "The temperature unit to use. Infer this from the user's location."}}, 'required': ['location', 'format']}}}], 'tool_choice': 'auto'}
